Separate simulator from main.cpp (#21)
* Separate Simulator

* Add newline

* Separate image creation

* Format

* Try to make headers visible in visual studio

* Hack: Make header files available in visual studio
